requirements.txt
tensorflow
torch
opencv-python
numpy
pillow
pandas
transformers
scikit-learn
matplotlib
pyqt5
fastapi
uvicorn
ultralytics
gradio
pyyaml
pytest
python-multipart
httpx
tensorboard